requirements.txt
pandas
sqlalchemy
virtualenv
dotenv
psycopg2-binary
openpyxl
statsmodels
matplotlib